pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "pre_commit_text"
description = "Some text hooks for pre-commit."
authors = [
  { name="Raphael Boidol", email="boidolr@users.noreply.github.com" },
]
urls = {"Respository" = "https://github.com/boidolr/pre-commit-text"}
license = {file = "LICENSE"}
readme = "README.md"
dynamic = ["version"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
]

requires-python = ">=3.10"
dependencies = [
    "ruamel.yaml==0.18.5",
]

[project.optional-dependencies]
dev = [
    "pre-commit",
    "ruff",
    "tox",
    "tox-pyenv",
]
tests = [
    "pytest",
    "pytest-clarity",
    "pytest-ruff",
    "pytest-mypy",
]

[project.scripts]
format-yaml = "pre_commit_text.format_yaml:main"
search-replace = "pre_commit_text.search_replace:main"
replace-tabs = "pre_commit_text.replace_tabs:main"

[tool.setuptools.dynamic]
version = {attr = "pre_commit_text.VERSION"}

[tool.setuptools]
packages = ["pre_commit_text"]

[tool.ruff]
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # pyupgrade
    "UP",
    # refurb
    "FURB",
    # flake8-bandit
    "S",
    # flake8-comprehensions
    "C4",
    # flake8-logging-format
    "G",
    # flake8-annotations
    "ANN",
    # flake8-simplify
    "SIM",
    # flake8-pytest-style
    "PT",
    # flake8-return
    "RET",
    # flake8-use-pathlib
    "PTH",
    # flake8-tidy-imports (relative-imports)
    "TID252",
    # Perflint
    "PERF",
    # flynt
    "FLY",
    # isort
    "I001",
    # Unneeded noqa
    "RUF100"
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.ruff.lint.flake8-annotations]
ignore-fully-untyped = true

[tool.ruff.lint.isort]
force-single-line = true
order-by-type = false

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "--mypy --ruff"
filterwarnings = [
    "error",
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
